[APP] Don't delete Zar extraction folder on failure if empty

+++ src/xenia/app/emulator_window.cc
@@ -1322,8 +1322,9 @@ void EmulatorWindow::ExtractZarchive() {
       if (result != X_STATUS_SUCCESS) {
         std::error_code ec;
 
-        // delete incomplete output file
-        std::filesystem::remove(abs_extract_dir, ec);
+        if (!std::filesystem::is_empty(abs_extract_dir)) {
+          std::filesystem::remove(abs_extract_dir, ec);
+        }
 
         summary += fmt::format("\nFailed: {}", zarchive_file_path);
